Fixed a bug
git-tfs-id: [https://devops.cyberoptics.com/TPC1]$/CyberStitch;C44122

+++ Main/src/logic/ManagedPanelAlignment/ManagedPanelAlignment.cpp
@@ -120,7 +120,7 @@ namespace PanelAlignM {
 		mSet->dOverallConfidence = resultsSet->CalConfidence();
 		// Confidence for each device
 		unsigned int iNumDevice = _pAligner->GetMosaicSet()->GetNumDevice();
-		mSet->dDeviceConfidences = gcnew array<double>(iNumDevice);
+		mSet->dDeviceConfidences = gcnew array<double>((int)iNumDevice);
 		for(int i=0; i<iNumDevice; i++)
 			mSet->dDeviceConfidences[i] = resultsSet->CalConfidence(i);
